requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
jinja2==3.1.5
sqlalchemy==2.0.38
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.6
email-validator==2.2.0
reportlab==4.3.1
python-docx==1.1.2
itsdangerous==2.2.0
